[REFACTOR] move FFid to algorithm
