+++ src/openms/include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/sources.cmake
@@ -21,6 +21,7 @@ FeatureFinderAlgorithmIsotopeWavelet.h
 FeatureFinderAlgorithmMRM.h
 FeatureFinderAlgorithmPicked.h
 FeatureFinderAlgorithmPickedHelperStructs.h
+FeatureFinderIdentificationAlgorithm.h
 FeatureFinderDefs.h
 Fitter1D.h
 GaussFitter1D.h
